Factor out `SequencerStep`
